Copy player logs only if they exist in runner.sh

+++ docker/runner/runner.sh
@@ -8,8 +8,15 @@ cd /root/simulator/bin
 ./main
 
 # Copy game log to output dir
-cp /root/simulator/bin/game.log /root/output_log
+if [ -f /root/simulator/bin/game.log  ]; then
+	cp /root/simulator/bin/game.log /root/output_log
+fi
 
 # Copy debug logs to output dir
-cp /root/simulator/bin/player_1.dlog /root/output_log
-cp /root/simulator/bin/player_2.dlog /root/output_log
+if [ -f /root/simulator/bin/player_1.dlog  ]; then
+	cp /root/simulator/bin/player_1.dlog /root/output_log
+fi
+
+if [ -f /root/simulator/bin/player_2.dlog  ]; then
+	cp /root/simulator/bin/player_2.dlog /root/output_log
+fi
